2012-09-14
Added interaction between bubble and block.

Changes:

Class: Game
- Added some new code in the calculate() method that iterates through
all the bubbles in the bubble pool and all the blocks in the block
pool and does a calculation + update where necessary.

Class: Calculate
- Added new interaction method calculateBubbleBlock() that calculates
the contextual interaction between bubble and block.
- If it finds that the bubble is within the x and y parameters of
touching the block, will return a boolean to the Game.calculate()
method which breaks out of updating the bubble movement.
- As you can see, the Calculate class will hold methods for
calculating every possible combination of objects. The contextual
relationships are defined here.

The collision rendering is incorrect, but I'll fix that later.