2012-09-23
Mainly added some changes to how the classes are handled and how the
rendering is handled.

Changes:

Class: GameLoop
- Moved the Render object into the Game class.
- Changed most static instances into dynamic instances.
- Added Game.delta field for smoother renders.

Class: Game
- Added instance of the Render.
- Changed most static instances into dynamic instances (main reason is
to not have as many "global" objects, to enforce accessing objects in
the best class).

Class: Calculate
- Added the Game.delta value to any calculation that has to do with
rate (distances).
- The delta value is calculated in the GameLoop to figure out any
timer disparities (since the Java timer is not guaranteed).
- The delta value, when factored into calculations, will place the
position of any object's movement in the correct position relative to
real time (as opposed to game frame time, like before).
- The delta value reduces jerky movement in rendering when the game's
FPS changes (since the value ties the calculation of position relative
to real time, and not FPS).

Class: View
- Added antialiasing for rendering.